refactor(api): per-provider packages under src/api/providers
Move Gemini and Ollama client code into src/api/providers/{gemini,ollama}
so additional providers can be added without piling siblings into a flat
directory. Rename the factory from GeminiClientFactory to ModelClientFactory
to match its actual responsibility (it dispatches to either provider based
on settings.provider).

Public SDK breaking change: GeminiClientFactory is renamed to
ModelClientFactory. OllamaClient and OllamaClientConfig are now exported
alongside the Gemini equivalents. Recorded in release-notes.json as the
4.9.0 stub entry.

Gemini-only services (image generation, deep research, Google Search, web
fetch) are unchanged and continue to use the @google/genai SDK directly;
a later refactor will introduce provider-agnostic interfaces for those.
