+++ src/api/providers/gemini/config.ts
@@ -1,0 +1,11 @@
+/**
+ * Configuration for GeminiClient
+ */
+export interface GeminiClientConfig {
+	apiKey: string;
+	model?: string;
+	temperature?: number;
+	topP?: number;
+	maxOutputTokens?: number;
+	streamingEnabled?: boolean;
+}
