+++ src/api/providers/gemini/index.ts
@@ -1,0 +1,2 @@
+export { GeminiClient } from './client';
+export type { GeminiClientConfig } from './config';
