+++ src/api/providers/ollama/config.ts
@@ -1,0 +1,8 @@
+export interface OllamaClientConfig {
+	baseUrl: string;
+	model?: string;
+	temperature?: number;
+	topP?: number;
+	maxOutputTokens?: number;
+	streamingEnabled?: boolean;
+}
