+++ src/api/providers/ollama/index.ts
@@ -1,0 +1,2 @@
+export { OllamaClient } from './client';
+export type { OllamaClientConfig } from './config';
