+++ src/services/scheduled-task-runner.ts
@@ -4,7 +4,7 @@ import type { ScheduledTask } from './scheduled-task-manager';
 import { ToolCategory, DestructiveAction } from '../types/agent';
 import type { ConfirmationResult, DiffContext, IConfirmationProvider, Tool } from '../tools/types';
 import { ToolExecutionContext } from '../tools/types';
-import { GeminiClientFactory } from '../api';
+import { ModelClientFactory } from '../api';
 import { ExtendedModelRequest } from '../api/interfaces/model-api';
 import { ensureFolderExists } from '../utils/file-utils';
 import { formatLocalDate, formatLocalTimestamp } from '../utils/format-utils';
@@ -85,7 +85,7 @@ export class ScheduledTaskRunner {
 		}
 
 		const toolContext: ToolExecutionContext = { plugin: this.plugin, session };
-		const modelApi = GeminiClientFactory.createChatModel(this.plugin);
+		const modelApi = ModelClientFactory.createChatModel(this.plugin);
 		const availableTools = this.plugin.toolRegistry.getEnabledTools(toolContext);
 
 		// Prepend a turn preamble so the model has accurate "now" awareness.
