+++ test/api/providers/gemini/client.test.ts
@@ -1,7 +1,8 @@
 import type { Mock } from 'vitest';
-import { GeminiClient, GeminiClientConfig } from '../../src/api/gemini-client';
-import { GeminiPrompts } from '../../src/prompts';
-import type { ExtendedModelRequest } from '../../src/api/interfaces/model-api';
+import { GeminiClient } from '../../../../src/api/providers/gemini/client';
+import type { GeminiClientConfig } from '../../../../src/api/providers/gemini/config';
+import { GeminiPrompts } from '../../../../src/prompts';
+import type { ExtendedModelRequest } from '../../../../src/api/interfaces/model-api';
 
 // Capture every call to `client.models.generateContent` so tests can assert on
 // the params (system instruction, contents, etc.) the SDK sees. vi.hoisted lets
